Lazily start orchestrator and improve test setup
This allows us to not start the orchestrator if we don't really need to.
