+++ package.json
@@ -11,7 +11,8 @@
     "start": "ts-node -r module-alias/register bin/start.ts",
     "watch": "ts-node -r module-alias/register bin/watch.ts",
     "lint": "eslint 'src/**/*.ts' 'bin/**/*.ts' 'agent/**/*.ts'",
-    "test": "mocha -r module-alias/register -r ts-node/register test/**/*.test.ts",
+    "test": "mocha -r ./test/setup test/**/*.test.ts",
+    "mocha": "mocha -r ./test/setup",
     "test:app:start": "cd test/app && react-scripts start",
     "test:app:build": "cd test/app && react-scripts build"
   },
