+++ test/setup.js
@@ -1,0 +1,7 @@
+require('module-alias/register');
+require('ts-node/register');
+process = require('process');
+
+let { setLogLevel } = require('../src/log-level')
+
+setLogLevel(process.env.LOG_LEVEL || 'warn');
